services/orchestrator/pyproject.toml
[project]
name = "apf-orchestrator"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.7",
    "pydantic-settings>=2.2",
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.20",
    "redis[asyncio]>=5.0",
    "networkx>=3.3",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "httpx>=0.27",
    "prometheus-client>=0.20",
    "structlog>=24.1",
]

[project.optional-dependencies]
dev = ["pytest>=8", "pytest-asyncio>=0.23", "pytest-cov>=5", "fakeredis[aioredis]>=2.23"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
